refactor: PR suggestions
