+++ packages/bull/src/domain/bull-queue.domain.ts
@@ -1,0 +1,7 @@
+import { BullQueueProcessor } from '@nestjs/bull'
+
+export type BullModuleQueue = {
+  name: string
+  configKey?: string
+  processors?: BullQueueProcessor[]
+}
