+++ packages/bull/src/domain/index.ts
@@ -1,2 +1,3 @@
-export * from './bull-module.options'
+export * from './bull-options.domain'
+export * from './bull-queue.domain'
 export * from './constants'
